refractor main.js
