--- electron/main.js
@@ -1,5 +1,0 @@
-import{app as D,BrowserWindow as Ne,ipcMain as x,dialog as ye,protocol as Te,shell as _e}from"electron";import{createRequire as Qe}from"node:module";import u from"node:path";import{fileURLToPath as Ye,pathToFileURL as vt}from"node:url";import Z from"node:https";import m from"node:fs";import Ze from"node:zlib";import{spawn as Q}from"node:child_process";import et from"node:os";import P from"node:fs";import b from"node:path";import je from"node:crypto";import{pipeline as ze}from"node:stream";import{promisify as We}from"node:util";import ie from"node:https";var Be=We(ze);function ae(o){return String(o||"").replace(/\\/g,"/").replace(/^\/+/,"")}function xe(o){return String(o||"").replace(/\\/g,"/").toLowerCase()}var K=new Map,Pe=new ie.Agent({keepAlive:!0,maxSockets:64,maxFreeSockets:32,keepAliveMsecs:6e4,scheduling:"fifo"});function H(o){return new Promise((t,r)=>{let e=je.createHash("sha256"),n=P.createReadStream(o);n.on("error",r),n.on("end",()=>t(e.digest("hex"))),n.on("data",s=>e.update(s))})}function Ce(o){P.mkdirSync(o,{recursive:!0})}function qe(o){return new Promise((t,r)=>{let e=ie.get(o,{agent:Pe},n=>{if(n.statusCode!==200){r(new Error(`HTTP ${n.statusCode} for ${o}`)),n.resume();return}let s="";n.setEncoding("utf8"),n.on("data",a=>s+=a),n.on("end",()=>{try{t(JSON.parse(s))}catch(a){r(a)}})});e.on("error",r),e.setTimeout(3e4,()=>{e.destroy(),r(new Error("JSON fetch timeout"))})})}function z(o){return new Promise(t=>setTimeout(t,o))}function Ae(){return{cancelled:!1,requests:new Set}}function ce(o){if(o){o.cancelled=!0;for(let t of o.requests)try{t.destroy(new Error("cancelled"))}catch{}o.requests.clear()}}var Oe=new Set(["ECONNRESET","ETIMEDOUT","EAI_AGAIN","ECONNABORTED","ENETRESET","EHOSTUNREACH","ENETUNREACH","EPIPE","ENOTFOUND","ECONNREFUSED","EHOSTDOWN","ENETDOWN"]);async function J(o,t,r,e=1,n,s=0,a=0){return new Promise((i,d)=>{Ce(b.dirname(t));let c=P.createWriteStream(t,{flags:s>0?"a":"w"});if(n?.cancelled)return c.close(()=>{}),d(new Error("cancelled"));let h={agent:Pe};s>0&&(h.headers={Range:`bytes=${s}-`});let f=ie.get(o,h,l=>{if(l.statusCode!==200&&!(s>0&&l.statusCode===206)){if(c.close(()=>{}),s>0&&l.statusCode===200){try{P.unlinkSync(t)}catch{}let y=Math.min(2e3*e,1e4);l.resume();try{f.destroy(new Error("restart:range-ignored"))}catch{}return z(y).then(()=>i(J(o,t,r,e+1,n,0,a)))}if(P.unlink(t,()=>{}),l.statusCode&&[429,500,502,503,504].includes(l.statusCode)&&e<5){let y=Math.min(2e3*e,1e4);return l.resume(),z(y).then(()=>i(J(o,t,r,e+1,n,s,a)))}d(new Error(`HTTP ${l.statusCode} for ${o}`)),l.resume();return}let g=Number(l.headers["content-length"]||0),S=a>0?a:s>0?s+g:g,p=0,E=Date.now(),w=e<=2?3e4:45e3,v=setInterval(()=>{if(!n?.cancelled&&Date.now()-E>w)try{let y=new Error("stalled");y.code="ETIMEDOUT",f.destroy(y)}catch{}},1e4);l.on("data",y=>{p+=y.length,E=Date.now(),r&&r(Math.min(s+p,S||s+p),S||0)}),l.on("aborted",async()=>{let y=new Promise(C=>c.close(()=>C()));try{clearInterval(v)}catch{}try{await y}catch{}if(e<5){let C=Math.min(2e3*e,1e4);return z(C).then(()=>i(J(o,t,r,e+1,n,s+p,a)))}d(new Error("response aborted"))}),Be(l,c).then(()=>{try{clearInterval(v)}catch{}i()}).catch(y=>{try{clearInterval(v)}catch{}d(y)})});if(n?.requests.add(f),f.on("socket",l=>{try{l.setKeepAlive(!0,6e4),l.setNoDelay(!0),l.setRecvBufferSize&&l.setRecvBufferSize(64*1024),l.setSendBufferSize&&l.setSendBufferSize(64*1024)}catch{}}),f.on("close",()=>n?.requests.delete(f)),f.setTimeout(9e4,()=>{try{let l=new Error("Request timeout");l.code="ETIMEDOUT",f.destroy(l)}catch{}}),n?.cancelled)try{f.destroy(new Error("cancelled"))}catch{}f.on("error",async l=>{let g=l?.code,S=e<8&&g&&Oe.has(String(g)),p=new Promise(E=>{try{c.close(()=>E())}catch{E()}});try{await p}catch{}if(S){let E=Math.min(1e3*Math.pow(1.5,e),15e3),w=Math.random()*1e3,v=E+w,y=s;try{y=P.statSync(t).size}catch{}return z(v).then(()=>i(J(o,t,r,e+1,n,y,a)))}d(l)})})}async function He(o,t,r){try{let e=P.statSync(o);return r&&Number(r)>0&&e.size!==Number(r)?!1:(await H(o)).toLowerCase()===String(t).toLowerCase()}catch{return!1}}async function be(o){let t=`${o.replace(/\/$/,"")}/checksums.json`;return qe(t)}async function Je(o,t,r,e,n=4,s){let a=ae(t.path).split("/").join(b.sep),i=b.join(r,a);if(Ce(b.dirname(i)),await He(i,t.checksum,t.size))return e("progress:skip",{path:t.path}),{downloadComplete:!0,verificationComplete:!0,targetPath:i};if(s?.cancelled)throw new Error("cancelled");if(Array.isArray(t.parts)&&t.parts.length>0){let c=t.parts.length,h=new Array(c),f=0,l=new Set;async function g(){for(;;){let p=f++;if(p>=c)return;if(l.has(p))continue;l.add(p);let E=t.parts[p],w=ae(E.path),v=`${o.replace(/\/$/,"")}/${w}`,y=`${i}.part${p}`;try{if(P.existsSync(y))if((await H(y)).toLowerCase()===String(E.checksum).toLowerCase()){e("progress:part",{path:t.path,part:p,totalParts:c,received:Number(E.size||0),total:Number(E.size||0)}),h[p]=y;continue}else try{P.unlinkSync(y)}catch{}}catch{}let C=0;for(;;){if(s?.cancelled)throw new Error("cancelled");C+=1;let A=0,$=0;try{let F=0;try{F=P.statSync(y).size}catch{}let N=Number(E.size||0);if(N>0&&F>N)try{P.unlinkSync(y),F=0}catch{}await J(v,y,(L,W)=>{let B=Math.max(0,L-$);$=L;try{B>0&&e("progress:bytes",{delta:B})}catch{}A+=B;let O=N||W||0;e("progress:part",{path:t.path,part:p,totalParts:c,received:Math.min(L,O||L),total:O})},1,s,F,N)}catch{try{P.unlinkSync(y)}catch{}try{A>0&&e("progress:bytes",{delta:-A})}catch{}try{e("progress:part:reset",{path:t.path,part:p,totalParts:c})}catch{}let N=Math.min(2e3*C,1e4);await z(N);continue}if((await H(y)).toLowerCase()===String(E.checksum).toLowerCase())break;try{P.unlinkSync(y)}catch{}try{A>0&&e("progress:bytes",{delta:-A})}catch{}try{e("progress:part:reset",{path:t.path,part:p,totalParts:c})}catch{}let T=Math.min(2e3*C,1e4);await z(T)}h[p]=y}}let S=Array.from({length:Math.max(1,n)},()=>g());return await Promise.all(S),{downloadComplete:!0,verificationComplete:!1,targetPath:i,mergeAndVerifyAsync:async()=>{e("progress:merge:start",{path:t.path,parts:h.length});let p=P.createWriteStream(i);for(let w=0;w<h.length;w++){let v=h[w];e("progress:merge:part",{path:t.path,part:w,totalParts:h.length}),await new Promise((y,C)=>{let A=P.createReadStream(v);A.on("error",C),p.on("error",C),A.on("end",()=>{try{P.unlinkSync(v)}catch{}y()}),A.pipe(p,{end:!1})})}await new Promise((w,v)=>{p.on("error",v),p.on("finish",w),p.end()});try{e("progress:merge:done",{path:t.path})}catch{}if(e("progress:verify",{path:t.path}),(await H(i)).toLowerCase()!==String(t.checksum).toLowerCase())throw new Error(`Checksum mismatch for ${t.path}`);try{let w=P.readdirSync(b.dirname(i)),v=b.basename(i);w.forEach(y=>{if(y.startsWith(v+".part"))try{P.unlinkSync(b.join(b.dirname(i),y))}catch{}})}catch{}return{verificationComplete:!0,targetPath:i,filePath:t.path}}}}else{let c=ae(t.path),h=`${o.replace(/\/$/,"")}/${c}`,f=0;for(;;){f+=1;let l=0,g=0,S=`${i}.download`,p=0;try{p=P.statSync(S).size}catch{}let E=Number(t.size||0);if(E>0&&p>E)try{P.unlinkSync(S),p=0}catch{}if(await J(h,S,(v,y)=>{let C=Math.max(0,v-g);g=v;try{C>0&&e("progress:bytes",{delta:C})}catch{}l+=C;let A=E||y||0;e("progress:file",{path:t.path,received:Math.min(v,A||v),total:A})},1,s,p,E),(await H(S)).toLowerCase()===String(t.checksum).toLowerCase())break;try{P.unlinkSync(S)}catch{}try{l>0&&e("progress:bytes",{delta:-l})}catch{}if(f>=2)throw new Error(`Checksum mismatch for ${t.path}`)}try{P.unlinkSync(i)}catch{}P.renameSync(`${i}.download`,i)}if(e("progress:verify",{path:t.path}),(await H(i)).toLowerCase()!==String(t.checksum).toLowerCase())throw new Error(`Checksum mismatch for ${t.path}`);return{downloadComplete:!0,verificationComplete:!0,targetPath:i}}async function De(o,t,r,e,n=!1,s=4,a=4,i,d){let c=new Set,h=[];for(let v of t.files||[]){if(!(n||!v.optional))continue;let y=xe(v.path);y&&(c.has(y)||(c.add(y),h.push(v)))}let f=h.filter(v=>!(Array.isArray(v.parts)&&v.parts.length>0)),l=h.filter(v=>Array.isArray(v.parts)&&v.parts.length>0),g=f.concat(l),S=g.length,p=0,E=g.reduce((v,y)=>{let C=Number(y.size||0);return C>0?v+C:Array.isArray(y.parts)&&y.parts.length?v+y.parts.reduce((A,$)=>A+Number($.size||0),0):v},0);try{e("progress:bytes:total",{totalBytes:E})}catch{}async function w(v,y,C,A=!1){let $=0,M=[];async function T(){for(;;){for(;d&&d();)if(await z(100),i?.cancelled)return;let N=$++;if(N>=v.length)return;let L=v[N],W=y+N,B=xe(L.path),O=async()=>Je(o,L,r,e,a,i),Se=!1,ee=K.get(B);ee?Se=!0:(e("progress:start",{index:W,total:S,path:L.path,completed:p}),ee=(async()=>{try{let j=await O();if(A&&j?.mergeAndVerifyAsync){let ve=j.mergeAndVerifyAsync().then(_=>(_?.filePath&&(p+=1,e("progress:done",{index:W,total:S,path:_.filePath,completed:p})),_)).catch(_=>{if(String(_?.message||_||"error").includes("cancelled"))throw _;try{let U=b.join(r,L.path.replace(/\\/g,b.sep));try{P.unlinkSync(U)}catch{}try{let R=P.readdirSync(b.dirname(U)),I=b.basename(U);R.forEach(te=>{if(te.startsWith(I+".part"))try{P.unlinkSync(b.join(b.dirname(U),te))}catch{}})}catch{}return O().then(R=>R.mergeAndVerifyAsync?R.mergeAndVerifyAsync().then(I=>(I?.filePath&&(p+=1,e("progress:done",{index:W,total:S,path:I.filePath,completed:p})),I)):R)}catch(U){try{e("progress:error",{path:L.path,message:String(U?.message||U)})}catch{}throw U}});return M.push(ve),{...j,skipProgressDone:!0}}return j}catch(j){if(String(j?.message||j||"error").includes("cancelled"))throw j;try{let _=b.join(r,L.path.replace(/\\/g,b.sep));try{P.unlinkSync(_)}catch{}let G=await O();if(A&&G?.mergeAndVerifyAsync){let U=G.mergeAndVerifyAsync().then(R=>(R?.filePath&&(p+=1,e("progress:done",{index:W,total:S,path:R.filePath,completed:p})),R)).catch(R=>{try{let I=b.join(r,L.path.replace(/\\/g,b.sep)),te=P.readdirSync(b.dirname(I)),Ie=b.basename(I);te.forEach(Ee=>{if(Ee.startsWith(Ie+".part"))try{P.unlinkSync(b.join(b.dirname(I),Ee))}catch{}})}catch{}try{e("progress:error",{path:L.path,message:String(R?.message||R)})}catch{}throw R});return M.push(U),{...G,skipProgressDone:!0}}return G}catch(_){try{e("progress:error",{path:L.path,message:String(_?.message||_)})}catch{}throw _}}})(),K.set(B,ee));let ke;try{ke=await ee}finally{Se||K.delete(B)}ke?.skipProgressDone||(p+=1,e("progress:done",{index:W,total:S,path:L.path,completed:p})),await z(10)}}let F=Array.from({length:Math.max(1,C)},()=>T());await Promise.all(F),M.length>0&&await Promise.all(M)}if(await w(f,0,Math.max(1,s)),K.size>0)try{await Promise.all([...K.values()])}catch{}await w(l,f.length,1,!0)}import le from"node:fs";import Ge from"node:path";import{app as Ke}from"electron";var $e=Ke.getPath("userData"),Le=Ge.join($e,"settings.json");function de(){try{let o=le.readFileSync(Le,"utf-8");return JSON.parse(o)}catch{return{}}}function Xe(o){le.mkdirSync($e,{recursive:!0}),le.writeFileSync(Le,JSON.stringify(o,null,2),"utf-8")}function he(o,t=void 0){return de()[o]??t}function re(o,t){let r=de();r[o]=t,Xe(r)}function ue(){return de()}var Ue=Qe(import.meta.url),{autoUpdater:V}=Ue("electron-updater"),Re=Ue("electron-log"),tt=Ye(import.meta.url),X=u.dirname(tt);var k,q=null,fe=new Set,ne=new Map,oe=new Map,pe="r5v",se=[];function Ve(o){try{return(o||[]).filter(t=>typeof t=="string"&&t.startsWith(`${pe}://`))}catch{return[]}}function ge(o){try{let t=new URL(o);if(t.hostname==="mod"&&t.pathname==="/install"){let r=t.searchParams.get("name")||"",e=t.searchParams.get("version")||"",n=t.searchParams.get("downloadUrl")?.split(",")||"";if(k&&k.webContents&&!k.webContents.isLoading())try{k.webContents.send("deeplink:mod-install",{url:o,name:r,version:e,downloadUrls:n})}catch{}else se.push(o)}}catch{}}function rt(){for(;se.length;){let o=se.shift();ge(o)}}var nt=D.requestSingleInstanceLock();nt?D.on("second-instance",(o,t)=>{try{Ve(t).forEach(e=>ge(e)),k&&(k.isMinimized()&&k.restore(),k.focus())}catch{}}):D.quit();D.on("open-url",(o,t)=>{o.preventDefault(),ge(t)});async function Me(){let o=he("windowState",{width:1150,height:800,x:void 0,y:void 0,isMaximized:!1});k=new Ne({width:o.width,height:o.height,x:o.x,y:o.y,minWidth:1150,minHeight:800,frame:!1,titleBarStyle:"hidden",webPreferences:{preload:u.join(X,"preload.cjs"),contextIsolation:!0,nodeIntegration:!1},show:!1,icon:u.join(X,"..","public","icon.png")});let t=process.env.VITE_DEV_SERVER_URL;if(t)await k.loadURL(t);else{let a=u.join(X,"..","dist","index.html");await k.loadFile(a)}k.webContents.setWindowOpenHandler(({url:a})=>{try{_e.openExternal(a)}catch{}return{action:"deny"}}),k.webContents.on("will-navigate",(a,i)=>{let d=i.startsWith("file:"),c=!!process.env.VITE_DEV_SERVER_URL&&i.startsWith(String(process.env.VITE_DEV_SERVER_URL));if(!d&&!c){a.preventDefault();try{_e.openExternal(i)}catch{}}});let r=()=>{k&&(k.isVisible()||k.show(),k.isFocused()||k.focus())};o.isMaximized&&k.maximize(),k.once("ready-to-show",r),k.webContents.once("did-finish-load",r);let e=setTimeout(r,1e3);k.on("show",()=>clearTimeout(e));function n(){if(!k)return;let a=k.getBounds(),i=k.isMaximized();re("windowState",{width:a.width,height:a.height,x:a.x,y:a.y,isMaximized:i})}k.on("resize",n),k.on("move",n),k.on("maximize",n),k.on("unmaximize",n),(!!process.env.VITE_DEV_SERVER_URL||process.env.NODE_ENV==="development")&&k.webContents.openDevTools({mode:"detach"}),k.webContents.on("did-fail-load",(a,i,d,c)=>{console.error("Load failed",{errorCode:i,errorDesc:d,validatedURL:c}),ye.showErrorBox("Load failed",`${d} (code ${i})
-URL: ${c}`),k.show()})}D.on("window-all-closed",()=>{process.platform!=="darwin"&&D.quit()});D.whenReady().then(()=>{try{process.defaultApp&&process.platform==="win32"?D.setAsDefaultProtocolClient(pe,process.execPath,[process.argv[1]]):D.setAsDefaultProtocolClient(pe)}catch{}let o=u.resolve(D.getAppPath(),".."),t=u.join(o,"cache");Te.registerFileProtocol("r5v",(e,n)=>{try{let s=new URL(e.url),a=decodeURIComponent(s.pathname).replace(/^\//,""),i=u.normalize(u.join(t,a));n({path:i})}catch{n({error:-6})}});let r=u.join(X,"..","dist");Te.interceptFileProtocol("file",(e,n)=>{try{let s=new URL(e.url),a=decodeURIComponent(s.pathname),i=a.replace(/\\/g,"/"),d=i.indexOf("/_astro/");if(d!==-1){let c=i.substring(d+8),h=u.join(r,"_astro",c);return n({path:h})}if(i.endsWith("/favicon.svg")||/\/favicon\.svg$/i.test(i)){let c=u.join(r,"favicon.svg");return n({path:c})}if(i.startsWith("/")){let c=i.replace(/^\/+/,""),h=u.join(r,c);try{return m.accessSync(h),n({path:h})}catch{}}return n({path:a})}catch{return n({error:-6})}}),Me();try{Ve(process.argv).forEach(e=>se.push(e))}catch{}try{k?.webContents?.once("did-finish-load",rt)}catch{}try{V.autoDownload=!1;try{V.logger=Re,Re.transports.file.level="info"}catch{}V.on("error",e=>{try{k?.webContents.send("update:error",{message:String(e?.stack||e?.message||e)})}catch{}}),V.on("update-available",e=>{try{k?.webContents.send("update:available",e)}catch{}}),V.on("update-not-available",e=>{try{k?.webContents.send("update:not-available",e)}catch{}}),V.on("download-progress",e=>{try{k?.webContents.send("update:download-progress",e)}catch{}}),V.on("update-downloaded",e=>{try{k?.webContents.send("update:downloaded",e)}catch{}})}catch{}D.on("activate",()=>{Ne.getAllWindows().length===0&&Me()})});x.handle("update:check",async()=>{try{return{ok:!0,result:(await V.checkForUpdates())?.updateInfo||null}}catch(o){return{ok:!1,error:String(o?.message||o)}}});x.handle("update:download",async()=>{try{return await V.downloadUpdate(),{ok:!0}}catch(o){return{ok:!1,error:String(o?.message||o)}}});x.handle("update:quitAndInstall",async()=>{try{return setImmediate(()=>V.quitAndInstall(!1,!0)),{ok:!0}}catch(o){return{ok:!1,error:String(o?.message||o)}}});x.handle("app:getVersion",async()=>{try{return D.getVersion()}catch{return"0.0.0"}});x.handle("app:getBaseDir",async()=>{try{return u.resolve(D.getAppPath(),"..")}catch{return""}});x.handle("app:getLauncherInstallRoot",async()=>{try{let o=process.env.LOCALAPPDATA||u.join(D.getPath("home"),"AppData","Local");return u.join(o,"Programs","r5vlauncher")}catch{return""}});function ot(o){try{let t=u.join(o,"mods.vdf"),r=m.readFileSync(t,"utf-8"),e={},n=/"([^"]+)"\s*"([01])"/g,s;for(;s=n.exec(r);){let a=s[1];a&&a!=="ModList"&&(e[a]=s[2]==="1")}return e}catch{return{}}}function we(o){try{let t=u.join(o,"mods.vdf"),r=m.readFileSync(t,"utf-8"),e={},n=[],s=/"([^"]+)"\s*"([01])"/g,a;for(;a=s.exec(r);){let i=a[1];!i||i==="ModList"||(i in e||n.push(i),e[i]=a[2]==="1")}return{map:e,order:n}}catch{return{map:{},order:[]}}}function me(o){try{let t=m.readFileSync(o,"utf-8"),r=t.match(/"id"\s*"([^"]+)"/i),e=t.match(/"name"\s*"([^"]+)"/i),n=r?r[1]:null,s=e?e[1]:null;return{id:n,name:s}}catch{return{id:null,name:null}}}function st(o,t){let r=['"ModList"',"{"];for(let[e,n]of Object.entries(t))r.push(`	"${e}"		"${n?"1":"0"}"`);r.push("}"),m.mkdirSync(o,{recursive:!0}),m.writeFileSync(u.join(o,"mods.vdf"),r.join(`
-`),"utf-8")}function Fe(o,t,r){let e=new Set,n=['"ModList"',"{"],s=a=>{n.push(`	"${a}"		"${r[a]?"1":"0"}"`),e.add(a)};for(let a of Array.isArray(t)?t:[])Object.prototype.hasOwnProperty.call(r,a)&&!e.has(a)&&s(a);for(let a of Object.keys(r))e.has(a)||s(a);n.push("}"),m.mkdirSync(o,{recursive:!0}),m.writeFileSync(u.join(o,"mods.vdf"),n.join(`
-`),"utf-8")}x.handle("mods:listInstalled",async(o,{installDir:t})=>{try{let r=u.join(t,"mods"),{map:e,order:n}=we(r),s=[],a=m.existsSync(r)?m.readdirSync(r,{withFileTypes:!0}):[];for(let d of a){if(!d.isDirectory())continue;let c=u.join(r,d.name),h=u.join(c,"manifest.json"),f=u.join(c,"mod.vdf"),l=u.join(c,"icon.png"),g=null;try{g=JSON.parse(m.readFileSync(h,"utf-8"))}catch{}let S=me(f),p=S.id||g?.name||d.name,E=null;try{E=`data:image/png;base64,${m.readFileSync(l).toString("base64")}`}catch{}s.push({id:p,name:g?.name||S.name||d.name,folder:d.name,version:g?.version_number||null,description:g?.description||"",enabled:p?!!e[p]:!1,hasManifest:m.existsSync(h),iconDataUrl:E})}let i=new Map(n.map((d,c)=>[d,c]));return s.sort((d,c)=>{let h=d.id!=null&&i.has(d.id)?i.get(d.id):Number.MAX_SAFE_INTEGER,f=c.id!=null&&i.has(c.id)?i.get(c.id):Number.MAX_SAFE_INTEGER;return h!==f?h-f:String(d.name||"").localeCompare(String(c.name||""))}),{ok:!0,mods:s}}catch(r){return{ok:!1,error:String(r?.message||r)}}});x.handle("mods:setEnabled",async(o,{installDir:t,name:r,enabled:e})=>{try{let n=u.join(t,"mods"),{map:s}=we(n);return s[r]=!!e,Fe(n,void 0,s),{ok:!0}}catch(n){return{ok:!1,error:String(n?.message||n)}}});x.handle("mods:reorder",async(o,{installDir:t,orderIds:r})=>{try{let e=u.join(t,"mods"),{map:n,order:s}=we(e),a=Array.isArray(r)?r.filter(h=>typeof h=="string"&&h&&Object.prototype.hasOwnProperty.call(n,h)):[],i=new Set(a),d=s.filter(h=>Object.prototype.hasOwnProperty.call(n,h)&&!i.has(h)),c=[...a,...d];return Fe(e,c,n),{ok:!0}}catch(e){return{ok:!1,error:String(e?.message||e)}}});x.handle("mods:uninstall",async(o,{installDir:t,folder:r})=>{try{let e=u.join(t,"mods"),n=u.join(e,r);return m.rmSync(n,{recursive:!0,force:!0}),{ok:!0}}catch(e){return{ok:!1,error:String(e?.message||e)}}});x.handle("mods:fetchAll",async(o,{query:t})=>{let r="https://thunderstore.io/c/r5valkyrie/api/v1/package-listing-index/",e=s=>new Promise((a,i)=>{Z.get(s,{headers:{"User-Agent":"R5Valkyrie-Launcher/1.0",Accept:"*/*"}},c=>{if(c.statusCode&&c.statusCode>=300&&c.statusCode<400&&c.headers.location){c.resume();let f=c.headers.location.startsWith("http")?c.headers.location:new URL(c.headers.location,s).toString();return a(e(f))}if(c.statusCode!==200)return c.resume(),i(new Error(`HTTP ${c.statusCode}`));let h=[];c.on("data",f=>h.push(Buffer.isBuffer(f)?f:Buffer.from(f))),c.on("end",()=>a(Buffer.concat(h)))}).on("error",i)}),n=s=>{try{return Ze.gunzipSync(s)}catch{return s}};try{let s=await e(r),a=JSON.parse(n(s).toString("utf8")),i=Array.isArray(a)?a:[],d=6,c=[],h=0;await Promise.all(Array.from({length:d}).map(async()=>{for(;h<i.length;){let l=h++,g=i[l];try{let S=await e(g),p=n(S).toString("utf8"),E=JSON.parse(p);Array.isArray(E)&&c.push(...E)}catch{}}}));let f=c;if(t&&String(t).trim()){let l=String(t).toLowerCase();f=f.filter(g=>String(g?.name||"").toLowerCase().includes(l)||String(g?.full_name||"").toLowerCase().includes(l))}return{ok:!0,mods:f}}catch(s){try{let i=await e("https://thunderstore.io/c/r5valkyrie/api/v1/package/"),d=JSON.parse(i.toString("utf8")),c=Array.isArray(d)?d:[];if(t&&String(t).trim()){let h=String(t).toLowerCase();c=c.filter(f=>String(f?.name||"").toLowerCase().includes(h)||String(f?.full_name||"").toLowerCase().includes(h))}return{ok:!0,mods:c}}catch(a){return{ok:!1,error:String(a?.message||s?.message||a||s)}}}});x.handle("mods:install",async(o,{installDir:t,name:r,downloadUrl:e})=>{if(!e.startsWith("https://thunderstore.io"))return{ok:!1,error:"Download URL does not point to Thunderstore"};try{let n=String(r||"").trim();if(!n)return{ok:!1,error:"Invalid mod name"};if(fe.has(n))return{ok:!0};fe.add(n);let s=u.join(t,"mods");m.mkdirSync(s,{recursive:!0});let a=u.join(s,`.__mod_${Date.now()}.zip`),i=(f,l=0)=>new Promise((g,S)=>{if(l>5)return S(new Error("Too many redirects"));let p=m.createWriteStream(a);Z.get(f,{headers:{"User-Agent":"R5Valkyrie-Launcher/1.0",Accept:"application/octet-stream,*/*;q=0.8",Referer:"https://thunderstore.io/",Connection:"keep-alive"}},w=>{if(w.statusCode&&w.statusCode>=300&&w.statusCode<400&&w.headers.location){w.resume(),p.close(()=>{});try{m.unlinkSync(a)}catch{}let $=w.headers.location.startsWith("http")?w.headers.location:new URL(w.headers.location,f).toString();return g(i($,l+1))}if(w.statusCode!==200){p.close(()=>{});try{m.unlinkSync(a)}catch{}return w.resume(),S(new Error(`HTTP ${w.statusCode}`))}let v=Number(w.headers["content-length"]||0),y=0,C=Date.now(),A=$=>{try{o?.sender?.send("mods:progress",{key:r,phase:$,received:y,total:v})}catch{}};w.on("data",$=>{y+=Buffer.isBuffer($)?$.length:Buffer.byteLength(String($));let M=Date.now();M-C>150&&(C=M,A("downloading"))}),w.pipe(p),p.on("finish",()=>{A("downloading"),p.close(g)})}).on("error",w=>{try{m.unlinkSync(a)}catch{}S(w)})});await i(e);try{o?.sender?.send("mods:progress",{key:r,phase:"extracting"})}catch{}let d=u.join(s,r);try{m.rmSync(d,{recursive:!0,force:!0})}catch{}m.mkdirSync(d,{recursive:!0}),await new Promise((f,l)=>{let g=Q("powershell.exe",["-NoProfile","-NonInteractive","-Command",`Expand-Archive -LiteralPath "${a}" -DestinationPath "${d}" -Force`],{stdio:"ignore"});g.on("exit",S=>S===0?f():l(new Error(`Expand-Archive failed ${S}`))),g.on("error",l)});try{m.unlinkSync(a)}catch{}let c=null;try{let f=u.join(d,"mod.vdf");if(m.existsSync(f))c=me(f).id;else{let l=m.readdirSync(d,{withFileTypes:!0});for(let g of l)if(g.isDirectory()){let S=u.join(d,g.name,"mod.vdf");if(m.existsSync(S)&&(c=me(S).id,c))break}}}catch{}let h=ot(s);h[c||r]=!0,st(s,h);try{o?.sender?.send("mods:progress",{key:r,phase:"done"})}catch{}return{ok:!0}}catch(n){return{ok:!1,error:String(n?.message||n)}}finally{try{fe.delete(String(r||""))}catch{}}});x.handle("mods:watch",async(o,{installDir:t})=>{try{let r=u.join(t,"mods");if(!m.existsSync(r))return{ok:!0};let e=r;if(ne.has(e))return{ok:!0};let n=m.watch(r,{persistent:!0},(s,a)=>{let i=e;clearTimeout(oe.get(i));let d=setTimeout(()=>{try{k?.webContents?.send("mods:changed",{installDir:t})}catch{}},300);oe.set(i,d)});return ne.set(e,n),{ok:!0}}catch(r){return{ok:!1,error:String(r?.message||r)}}});x.handle("mods:unwatch",async(o,{installDir:t})=>{try{let e=u.join(t,"mods"),n=ne.get(e);if(n){try{n.close()}catch{}ne.delete(e)}let s=oe.get(e);return s&&(clearTimeout(s),oe.delete(e)),{ok:!0}}catch(r){return{ok:!1,error:String(r?.message||r)}}});x.handle("mods:iconDataUrl",async(o,{installDir:t,folder:r})=>{try{let e=u.join(t,"mods",r,"icon.png");return await m.promises.access(e),{ok:!0,dataUrl:`data:image/png;base64,${(await m.promises.readFile(e)).toString("base64")}`}}catch(e){return{ok:!1,error:String(e?.message||e)}}});x.handle("game:listProtonVersions",async()=>{if(process.platform!=="linux")return{ok:!0,versions:[]};try{let o=[],t=D.getPath("home"),r=[u.join(t,".local","share","Steam","steamapps","common")],e=[u.join(t,".local","share","Steam","compatibilitytools.d")];for(let n of r)try{if(!m.existsSync(n))continue;let s=m.readdirSync(n,{withFileTypes:!0});for(let a of s){if(!a.isDirectory())continue;let i=a.name,d=u.join(n,i);if(i.toLowerCase().includes("proton")){let c=u.join(d,"files","bin","wineserver"),h=u.join(d,"dist","bin","wineserver");(m.existsSync(c)||m.existsSync(h))&&o.push({name:i,path:d})}}}catch(s){console.warn(`Failed to scan ${n}:`,s)}for(let n of e)try{if(!m.existsSync(n))continue;let s=m.readdirSync(n,{withFileTypes:!0});for(let a of s){if(!a.isDirectory())continue;let i=a.name,d=u.join(n,i);o.push({name:i,path:d})}}catch(s){console.warn(`Failed to scan ${n}:`,s)}return o.sort((n,s)=>s.name.localeCompare(n.name)),o.unshift({name:"Default (Latest UMU-Proton)",path:""}),{ok:!0,versions:o}}catch(o){return{ok:!1,error:String(o?.message||o),versions:[]}}});x.handle("select-directory",async()=>{let o=await ye.showOpenDialog({properties:["openDirectory","createDirectory"]});return o.canceled||o.filePaths.length===0?null:o.filePaths[0]});x.handle("settings:get",()=>ue());x.handle("settings:set",(o,{key:t,value:r})=>re(t,r));x.handle("download:checksums",async(o,{baseUrl:t})=>be(t));x.handle("download:all",async(o,{baseUrl:t,checksums:r,installDir:e,includeOptional:n,concurrency:s,partConcurrency:a,channelName:i,mode:d})=>{let c=(h,f)=>o.sender.send(h,f);try{if(!(String(d||"install").toLowerCase()==="install")){let f=u.join(e,"mods","mods.vdf");if(m.existsSync(f)){let g=(Array.isArray(r?.files)?r.files:[]).filter(S=>{try{return String(S?.path||S?.name||"").replace(/\\/g,"/").toLowerCase()!=="mods/mods.vdf"}catch{return!0}});r={...r||{},files:g}}}}catch{}q=Ae(),Y=!1;try{await De(t,r,e,c,!!n,Number(s)||4,Number(a)||4,q,()=>Y);try{let h=he("channels",{})||{};h[String(i||"default")]={installDir:e,gameVersion:r?.game_version||null,gameBaseUrl:t,lastUpdatedAt:Date.now()},re("channels",h)}catch(h){console.error("Failed to persist channel info",h)}}finally{q=null}return!0});x.handle("default-install-dir",(o,{channelName:t})=>{let r="";if(process.platform==="win32"){let e=process.env.LOCALAPPDATA||u.join(D.getPath("home"),"AppData","Local");r=u.join(e,"Programs","R5VLibrary")}else if(process.platform==="linux"){let e=u.join(D.getPath("home"),"Games");r=u.join(e,"R5VLibrary")}return t?u.join(r,t):r});x.handle("launcher:config",async(o,{url:t})=>(e=>new Promise((n,s)=>{Z.get(e,a=>{if(a.statusCode!==200){s(new Error(`HTTP ${a.statusCode} for ${e}`)),a.resume();return}let i="";a.setEncoding("utf8"),a.on("data",d=>i+=d),a.on("end",()=>{try{n(JSON.parse(i))}catch(d){s(d)}})}).on("error",s)}))(t));x.handle("eula:get",async()=>{let o="https://playvalkyrie.org/api/eula",t=r=>new Promise((e,n)=>{Z.get(r,s=>{if(s.statusCode!==200){n(new Error(`HTTP ${s.statusCode}`)),s.resume();return}let a=[];s.on("data",i=>a.push(Buffer.isBuffer(i)?i:Buffer.from(i))),s.on("end",()=>{try{let i=Buffer.concat(a).toString("utf8");e(JSON.parse(i))}catch(i){n(i)}})}).on("error",n)});try{return{ok:!0,json:await t(o)}}catch(r){return{ok:!1,error:String(r?.message||r)}}});x.handle("video:cache",async(o,{filename:t})=>{try{let e=`https://blaze.playvalkyrie.org/video_backgrounds/${t}`,n=u.join(u.resolve(D.getAppPath(),".."),"cache"),s=u.join(n,"videos"),a=u.join(s,t),i=u.join(X,"..","dist"),d=u.join(i,t);await m.promises.mkdir(s,{recursive:!0});try{if((await m.promises.stat(a)).size>0){try{await m.promises.access(d)}catch{await m.promises.copyFile(a,d)}return`./${t}`}else await m.promises.unlink(a)}catch{}return await new Promise((c,h)=>{let f=m.createWriteStream(a),l=Z.get(e,g=>{if(g.statusCode!==200){f.close(()=>{}),m.unlink(a,()=>{}),h(new Error(`HTTP ${g.statusCode} for video ${t}`)),g.resume();return}g.pipe(f),f.on("finish",()=>f.close(c)),f.on("error",S=>{m.unlink(a,()=>h(S))})});l.on("error",g=>{f.close(()=>{}),m.unlink(a,()=>{}),h(g)}),l.setTimeout(3e4,()=>{l.destroy(),f.close(()=>{}),m.unlink(a,()=>{}),h(new Error(`Video download timeout for ${t}`))})}),await m.promises.copyFile(a,d),`./${t}`}catch(r){throw console.error(`Video cache error for ${t}:`,r),r}});x.handle("fs:is-installed-in-dir",async(o,{path:t})=>{let r=!1,e=!1;try{await m.promises.access(u.join(t,"r5apex.exe")),r=!0}catch{}try{await m.promises.access(u.join(t,"r5apex_ds.exe")),e=!0}catch{}return r||e});var Y=!1;x.handle("download:pause",async()=>{Y=!0;try{k?.webContents.send("progress:paused",{})}catch{}return!0});x.handle("download:resume",async()=>{Y=!1;try{k?.webContents.send("progress:resumed",{})}catch{}return!0});x.handle("download:cancel",async()=>{try{ce(q)}catch{}q=null,Y=!1;try{k?.webContents.send("progress:cancelled",{})}catch{}return!0});x.handle("select-file",async(o,{filters:t})=>{let r=await ye.showOpenDialog({properties:["openFile"],filters:Array.isArray(t)?t:void 0});return r.canceled||r.filePaths.length===0?null:r.filePaths[0]});x.handle("game:launch",async(o,{channelName:t,installDir:r,mode:e,argsString:n,winePrefix:s,protonVersion:a})=>{try{if(!r)throw new Error("Missing installDir");let i=String(e).toUpperCase()==="SERVER"?"r5apex_ds.exe":"r5apex.exe",d=u.join(r,i);await m.promises.access(d).catch(()=>{throw new Error(`Executable not found: ${d}`)});let h=(l=>!l||typeof l!="string"?[]:(l.match(/(?:[^\s\"]+|\"[^\"]*\")+/g)||[]).map(S=>S.replace(/^\"|\"$/g,"")))(n),f;if(process.platform==="win32")f=Q(d,h,{cwd:r,detached:!0,stdio:"ignore",env:{...process.env}});else if(process.platform==="linux"){let l=s||u.join(D.getPath("home"),"Games","R5VLibrary","wineprefix");h.unshift(d);let g=u.join(D.getPath("home"),".local","share","Steam"),S=u.join(l,"drive_c","Program Files (x86)","Steam");m.mkdirSync(S,{recursive:!0}),["GameOverlayRenderer64.dll","steamclient64.dll","steamclient.dll"].forEach(E=>{let w=u.join(g,E),v=u.join(S,E);if(!m.existsSync(w))throw new Error(`Steam is not installed correctly!
-Source file not found: ${w}`);m.existsSync(v)?console.log(`Skipped (exists): ${E}`):(m.copyFileSync(w,v),console.log(`Copied: ${E}`))}),f=Q("umu-run",h,{cwd:r,detached:!0,stdio:"ignore",env:{...process.env,WINEPREFIX:l,PROTONPATH:a||void 0}})}return f.unref(),{ok:!0}}catch(i){return{ok:!1,error:String(i?.message||i)}}});x.handle("fix-folder-permissions",async(o,{selectedChannel:t})=>{let r=[],e=ue().channels[t].installDir;try{if(!e)return{ok:!1,error:"No folder path provided"};let n=u.resolve(e).replace(/\//g,"\\");if(!n||n.length<3)return{ok:!1,error:`Invalid path format: ${e}`};let s=n.length>=3&&n[1]===":"?n.slice(3):n;if(/[<>:"|?*]/.test(s))return{ok:!1,error:`Path contains invalid characters: ${n}`};let d=et.userInfo().username,c=u.join(process.resourcesPath,"elevate.exe"),h=m.existsSync(c);try{m.mkdirSync(n,{recursive:!0})}catch{}let f=[{exe:"icacls",args:[n,"/grant",`${d}:F`,"/t","/q"],desc:`Grant permissions to current user (${d})`},{exe:"icacls",args:[n,"/grant","Everyone:F","/t","/q"],desc:"Grant Everyone full permissions"},{exe:"icacls",args:[n,"/grant","Users:F","/t","/q"],desc:"Grant Users full permissions"}];for(let l=0;l<f.length;l++){let{exe:g,args:S,desc:p}=f[l];try{let E=await new Promise((w,v)=>{let y,C,A;h?(C=[c,["-wait",g,...S]],A={stdio:["ignore","pipe","pipe"],windowsHide:!0}):(C=[g,S],A={stdio:["ignore","pipe","pipe"],windowsHide:!0});try{y=Q(C[0],C[1],A)}catch(T){console.error("[Permission Fix] Spawn failed:",T),v(T);return}let $="",M="";y.stdout&&y.stdout.on("data",T=>{$+=T.toString()}),y.stderr&&y.stderr.on("data",T=>{M+=T.toString()}),y.on("exit",(T,F)=>{w({code:T,stdout:$,stderr:M,signal:F})}),y.on("error",T=>{v(T)}),setTimeout(()=>{try{y.kill()}catch{}v(new Error("Command timeout"))},3e4)});if(E.code!==0){let w=E.stderr||E.stdout||"Unknown error";if(!(w.includes("duplicate")||w.includes("already")||E.code===1332)){let y=`${p} failed (code ${E.code}): ${w}`;r.push(y)}}}catch(E){let w=`${p} error: ${E.message}`;r.push(w)}}try{let l=u.join(n,"test_write.tmp");m.writeFileSync(l,"test"),m.unlinkSync(l)}catch(l){let g=`Write test failed: ${l.message}`;r.push(g);try{if(m.mkdirSync(n,{recursive:!0}),(await new Promise(p=>{let E=Q("cmd",["/c",`icacls "${n}" /grant "%USERNAME%":F /q`],{stdio:"ignore",windowsHide:!0});E.on("exit",w=>p({code:w})),E.on("error",()=>p({code:1})),setTimeout(()=>{try{E.kill()}catch{}p({code:1})},5e3)})).code===0){let p=u.join(n,"test_write2.tmp");return m.writeFileSync(p,"test"),m.unlinkSync(p),{ok:!0,warnings:r}}}catch{}return{ok:!1,error:g,details:r}}return r.length>0?{ok:!0,warnings:r}:{ok:!0}}catch(n){let s=`Unexpected error: ${n?.message||n}`;return r.push(s),{ok:!1,error:s,details:r}}});x.on("window:minimize",()=>{k?.minimize()});x.on("window:maximize",()=>{k&&(k.isMaximized()?k.unmaximize():k.maximize())});x.on("window:close",()=>{k?.close()});D.on("before-quit",()=>{if(q)try{ce(q)}catch{}});
