--- electron/preload.cjs
@@ -1,1 +1,0 @@
-"use strict";var{contextBridge:l,ipcRenderer:o}=require("electron");l.exposeInMainWorld("electronAPI",{selectDirectory:()=>o.invoke("select-directory"),getSettings:()=>o.invoke("settings:get"),setSetting:(e,n)=>o.invoke("settings:set",{key:e,value:n}),fetchChecksums:e=>o.invoke("download:checksums",{baseUrl:e}),downloadAll:e=>o.invoke("download:all",e),onProgress:(e,n)=>o.on(e,(i,s)=>n(s)),getDefaultInstallDir:e=>o.invoke("default-install-dir",{channelName:e}),fetchLauncherConfig:e=>o.invoke("launcher:config",{url:e}),cacheBackgroundVideo:e=>o.invoke("video:cache",{filename:e}),isInstalledInDir:e=>o.invoke("fs:is-installed-in-dir",{path:e}),openExternal:e=>o.invoke("open-external",{url:e}),pauseDownload:()=>o.invoke("download:pause"),resumeDownload:()=>o.invoke("download:resume"),cancelDownload:()=>o.invoke("download:cancel"),selectFile:e=>o.invoke("select-file",{filters:e}),launchGame:e=>o.invoke("game:launch",e),listProtonVersions:()=>o.invoke("game:listProtonVersions"),checkForUpdates:()=>o.invoke("update:check"),downloadUpdate:()=>o.invoke("update:download"),quitAndInstall:()=>o.invoke("update:quitAndInstall"),getAppVersion:()=>o.invoke("app:getVersion"),getBaseDir:()=>o.invoke("app:getBaseDir"),getLauncherInstallRoot:()=>o.invoke("app:getLauncherInstallRoot"),fetchEula:()=>o.invoke("eula:get"),onUpdate:(e,n)=>o.on(e,(i,s)=>n(s)),minimize:()=>o.send("window:minimize"),maximize:()=>o.send("window:maximize"),close:()=>o.send("window:close"),listInstalledMods:e=>o.invoke("mods:listInstalled",{installDir:e}),setModEnabled:(e,n,i)=>o.invoke("mods:setEnabled",{installDir:e,name:n,enabled:i}),reorderMods:(e,n)=>o.invoke("mods:reorder",{installDir:e,orderIds:n}),uninstallMod:(e,n)=>o.invoke("mods:uninstall",{installDir:e,folder:n}),fetchAllMods:e=>o.invoke("mods:fetchAll",{query:e}),installMod:(e,n,i)=>o.invoke("mods:install",{installDir:e,name:n,downloadUrl:i}),onModsProgress:e=>o.on("mods:progress",(n,i)=>e(i)),getModIconDataUrl:(e,n)=>o.invoke("mods:iconDataUrl",{installDir:e,folder:n}),watchMods:e=>o.invoke("mods:watch",{installDir:e}),unwatchMods:e=>o.invoke("mods:unwatch",{installDir:e}),onModsChanged:e=>o.on("mods:changed",(n,i)=>e(i)),fixFolderPermissions:e=>o.invoke("fix-folder-permissions",e)});
