--- electron/services/store.js
@@ -1,1 +1,0 @@
-import n from"node:fs";import c from"node:path";import{app as u}from"electron";var i=u.getPath("userData"),s=c.join(i,"settings.json");function o(){try{let t=n.readFileSync(s,"utf-8");return JSON.parse(t)}catch{return{}}}function f(t){n.mkdirSync(i,{recursive:!0}),n.writeFileSync(s,JSON.stringify(t,null,2),"utf-8")}function g(t,r=void 0){return o()[t]??r}function m(t,r){let e=o();e[t]=r,f(e)}function l(){return o()}export{l as getAllSettings,g as getSetting,m as setSetting};
